+++ components/Goal/goals.styles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  listItemContainer: {
+    marginBottom: 8,
+
+    paddingVertical: 16,
+    paddingHorizontal: 32,
+
+    backgroundColor: '#f7cb94',
+  },
+  listItem: {
+    color: '#1a0800',
+    fontSize: 24,
+    textAlign: 'center',
+  },
+});
